feature: adiciona tratativa de erros correta na service de Usuario

+++ src/domains/Usuario/services/serviceUsuario.ts
@@ -1,11 +1,31 @@
 import prisma from '../../../../config/prismaClient.js';
 import bcrypt from 'bcrypt';
-import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
 import { Prisma } from '@prisma/client';
+import {QueryError, InvalidParamError} from '../../../../errors'
 
 class ServiceUsuario {
     // Cria novo usuário
     async criarUsuario(body: Prisma.UsuarioCreateInput) {
+
+        const usuarioExistente = await prisma.usuario.findUnique({
+            where:{
+                email: body.email
+            }
+        })
+
+        if (usuarioExistente){
+            throw new QueryError("Email já cadastrado!")
+        }
+        if (body.email == null){
+            throw new InvalidParamError("Email não informado!")
+        }
+        if (body.senha == null){
+            throw new InvalidParamError("Senha não informada!")
+        }
+        if (body.nome == null){
+            throw new InvalidParamError("Nome não informado!")
+        }
+
         const hashSenha = await bcrypt.hash(body.senha, 10);
         const usuario: Prisma.UsuarioCreateInput = {
             nome: body.nome,
@@ -25,18 +45,23 @@ class ServiceUsuario {
 
             return usuarioCriado;
         } catch (erro) {
-            if (erro instanceof PrismaClientKnownRequestError) {
-                if (erro.code === 'P2002') {
-                    throw new Error('Erro ao criar usuário: e-mail já está em uso');
-                }
-            }
-
             throw new Error(`Erro ao criar usuário: ${erro}`);
         }
     }
 
     // Deleta usuário com id especifico
     async deletarUsuario(id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             await prisma.usuario.delete({
                 where: {
@@ -50,6 +75,9 @@ class ServiceUsuario {
 
     // Lista todos os usuários registrados
     async listarUsuarios(limit?: number, order?: Prisma.SortOrder) {
+        if ((order != null) && (order != 'asc' || 'desc')){
+            throw new InvalidParamError("Sort order inválida! Ordens válidas: 'asc' para ordenamento ascendente e 'desc' para ordenamento descendente.")
+        }
         try {
             const usuarios = await prisma.usuario.findMany({
                 orderBy: {
@@ -68,6 +96,17 @@ class ServiceUsuario {
 
     // Retoma o usuário com id especificado
     async listarUsuarioID(id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             const usuario = await prisma.usuario.findUnique({
                 where: {
@@ -85,6 +124,17 @@ class ServiceUsuario {
 
     // Retoma o usuário com email especificado
     async listarUsuarioEmail(email: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                email: email
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Email inválido!")
+        }
+        if (email == null){
+            throw new InvalidParamError("Email não informado!")
+        }
         try {
             const usuario = await prisma.usuario.findUnique({
                 where: {
@@ -102,7 +152,20 @@ class ServiceUsuario {
 
     // Atualiza informações do usuário de id especificado
     async atualizaUsuario(id: string, body: Prisma.UsuarioUpdateInput) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
+
         const usuarioUpdate: Prisma.UsuarioUpdateInput = { ...body };
+
         if (typeof body.senha === 'string') {
             usuarioUpdate.senha = await bcrypt.hash(body.senha, 10);
         }
@@ -123,6 +186,17 @@ class ServiceUsuario {
     }
 
     async listaHistoricoUsuario(id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             const historico = await prisma.logMusica.findMany({
                 where: {
@@ -144,6 +218,17 @@ class ServiceUsuario {
     }
 
     async listaMusicasSalvasUsuario(id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id inválido!")
+        }
+        if (id == null){
+            throw new InvalidParamError("Id não informado!")
+        }
         try {
             const musicasSalvas = await prisma.musicaSalva.findMany({
                 where: {
@@ -174,6 +259,30 @@ class ServiceUsuario {
     }
 
     async salvaMusicaUsuario(usuario_id: string, musica_id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: usuario_id
+            }
+        })
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: musica_id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id de usuário inválido!")
+        }
+        if (usuario_id == null){
+            throw new InvalidParamError("Id de usuário não informado!")
+        }
+
+        if (musicaEncontrada == null){
+            throw new QueryError("Id de música inválido!")
+        }
+        if (musica_id == null){
+            throw new InvalidParamError("Id de música não informado!")
+        }
+
         try {
             const musicaSalva = await prisma.musicaSalva.create({
                 data: {
@@ -196,6 +305,29 @@ class ServiceUsuario {
     }
 
     async removeMusicaSalvaUsuario(usuario_id: string, musica_id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: usuario_id
+            }
+        })
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: musica_id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id de usuário inválido!")
+        }
+        if (usuario_id == null){
+            throw new InvalidParamError("Id de usuário não informado!")
+        }
+
+        if (musicaEncontrada == null){
+            throw new QueryError("Id de música inválido!")
+        }
+        if (musica_id == null){
+            throw new InvalidParamError("Id de música não informado!")
+        }
         try {
             await prisma.musicaSalva.delete({
                 where: {
@@ -211,6 +343,29 @@ class ServiceUsuario {
     }
 
     async criaHistoricoUsuario(usuario_id: string, musica_id: string) {
+        const usuarioEncontrado = await prisma.usuario.findUnique({
+            where:{
+                id: usuario_id
+            }
+        })
+        const musicaEncontrada = await prisma.musica.findUnique({
+            where:{
+                id: musica_id
+            }
+        })
+        if (usuarioEncontrado == null){
+            throw new QueryError("Id de usuário inválido!")
+        }
+        if (usuario_id == null){
+            throw new InvalidParamError("Id de usuário não informado!")
+        }
+
+        if (musicaEncontrada == null){
+            throw new QueryError("Id de música inválido!")
+        }
+        if (musica_id == null){
+            throw new InvalidParamError("Id de música não informado!")
+        }
         try {
             const logMusica = await prisma.logMusica.create({
                 data: {
@@ -233,6 +388,18 @@ class ServiceUsuario {
     }
 
     async removeHistoricoUsuario(log_id: string) {
+        const logEncontrado = await prisma.logMusica.findUnique({
+            where:{
+                id: log_id
+            }
+        })
+
+        if (logEncontrado == null){
+            throw new QueryError("Id de log inválido!")
+        }
+        if (log_id == null){
+            throw new InvalidParamError("Id de log não informado!")
+        }
         try {
             await prisma.logMusica.delete({
                 where: {
